Repeat message processing on race condition
If another processor inserts a subject as this transaction
processes, we could end up with a race condition on commit.

This rescues the exception and retries message processing.

We limit ourselves to 10 cycles to avoid the risk of an unidentified
but causing the event processor to become locked in an infinite loop.
(For example, if we accidentally removed validate uniqueness of event
UUID we'd end up with the same situation, but it would never resolve.)