Remove swagger-diff gem
swagger-diff is used to analyse the differences between two
swagger schema. It can be used by the rake tasks for the unmaintained
jsonapi_swagger_helpers to check for backwards compatibility.

https://jsonapi-suite.github.io/jsonapi_suite_deprecated/ruby/backwards-compatibility

However it locks us to ruby 2.7, and generally its not a good idea
to keep unmaintained dependencies around. Given we're not using
any of this functionality I've pruned this back for now.